GVT-1718 Refactor folder structure and naming for geometry plan in Geoviite design lib
