+++ ui/src/geoviite-design-lib/geometry-plan/plan-decision-phase.tsx
