+++ ui/src/geoviite-design-lib/geometry-plan/plan-phase.tsx
