+++ ui/src/tool-panel/geometry-plan-infobox.tsx
@@ -7,8 +7,8 @@ import InfoboxField from 'tool-panel/infobox/infobox-field';
 import { GeometryPlanHeader } from 'geometry/geometry-model';
 import { useTranslation } from 'react-i18next';
 import { formatDateShort, toDateOrUndefined } from 'utils/date-utils';
-import PlanPhase from 'geoviite-design-lib/plan-phase/plan-phase';
-import PlanDecisionPhase from 'geoviite-design-lib/plan-decision/plan-decision-phase';
+import PlanPhase from 'geoviite-design-lib/geometry-plan/plan-phase';
+import PlanDecisionPhase from 'geoviite-design-lib/geometry-plan/plan-decision-phase';
 import { TrackNumberLink } from 'geoviite-design-lib/track-number/track-number-link';
 import { differenceInYears } from 'date-fns';
 import InfoboxButtons from 'tool-panel/infobox/infobox-buttons';
